grass.script: Create new location without a session (#3432)
The function create_location now works without a full session. Internally, it sets up a runtime environment to execute tools (modules) without the connection to a location (project) which is sufficient for executing g.proj to create a location (project).

The new test covers the new functionality, but to actually test it, the test needs to run by itself because grass.script.setup.init (still) creates a global session.

env is always modified to avoid confusion when it is modified and when it is not (as opposed to the original idea of accommodating usage with os.environ by modifying that). Copy is up to the user which makes it clear when that actually happens which is longer but more readable.